feat: create useSearchParamsUpdater hook
